Course detail page improvements and bug fixes

+++ apps/web/models/course.py
@@ -99,8 +99,8 @@ class Course(models.Model):
     # subnumber = models.IntegerField(null=True, db_index=True, blank=True)
     # source = models.CharField(max_length=16, choices=SOURCES.CHOICES)
 
-    difficulty_score = models.IntegerField(default=0)
-    quality_score = models.IntegerField(default=0)
+    difficulty_score = models.FloatField(default=0.0)
+    quality_score = models.FloatField(default=0.0)
 
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
